Rename "physics" to "colliders"
Because that's all that's left after "ground_detection" is split off. (Also technically there's no physics calculations being done here.)
