+++ examples/platformer/colliders.rs
