+++ examples/platformer/misc_objects.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 use bevy_ecs_ldtk::prelude::*;
 
-use crate::physics::ColliderBundle;
+use crate::colliders::ColliderBundle;
 
 #[derive(Clone, Default, Bundle, LdtkEntity)]
 pub struct ChestBundle {
